+++ simson/__init__.py
@@ -1,0 +1,3 @@
+from simson import plastics
+from simson import steel
+from simson import common
